Trait str_utils::StrUtilsExt

pub trait StrUtilsExt {

    fn ends_with_ai(self, pat: &str) -> bool
    where
        Self: Sized;

    fn ends_with_ai_ci(self, pat: &str) -> bool
    where
        Self: Sized;

    fn ends_with_ci(self, pat: &str) -> bool
    where
        Self: Sized;

    fn find_ai(self, pat: &str) -> Option<usize>
    where
        Self: Sized;

    fn find_ai_ci(self, pat: &str) -> Option<usize>
    where
        Self: Sized;

    fn find_ci(self, pat: &str) -> Option<usize>
    where
        Self: Sized;

    fn starts_with_ai(self, pat: &str) -> bool
    where
        Self: Sized;

    fn starts_with_ai_ci(self, pat: &str) -> bool
    where
        Self: Sized;

    fn starts_with_ci(self, pat: &str) -> bool
    where
        Self: Sized;

    fn no_accent(self) -> String
    where
        Self: Sized;

    fn no_accent_lowercase(self) -> String
    where
        Self: Sized;

    fn no_accent_uppercase(self) -> String
    where
        Self: Sized;

    fn contains_ai(self, pat: &str) -> bool
    where
        Self: Sized,
    { ... }

    fn contains_ai_ci(self, pat: &str) -> bool
    where
        Self: Sized,
    { ... }

    fn contains_ci(self, pat: &str) -> bool
    where
        Self: Sized,
    { ... }

}

Trait for searching string with accent / case insensitive comparison.

Required methods

fn ends_with_ai(self, pat: &str) -> bool where
Self: Sized,

Returns true if the given pattern matches a suffix of this string slice. Returns false if it does not.

The comparison is accent insensitive.

Example

use str_utils::StrUtilsExt;

assert!("Press Café".ends_with_ai("Cafe"));

fn ends_with_ai_ci(self, pat: &str) -> bool where
Self: Sized,

Returns true if the given pattern matches a suffix of this string slice. Returns false if it does not.

The comparison is accent / case insensitive.

Example

use str_utils::StrUtilsExt;

assert!("Press Café".ends_with_ai_ci("cafe"));

fn ends_with_ci(self, pat: &str) -> bool where
Self: Sized,

Returns true if the given pattern matches a suffix of this string slice. Returns false if it does not.

The comparison is case insensitive.

Example

use str_utils::StrUtilsExt;

assert!("Press Cafe".ends_with_ci("cafe"));

fn find_ai(self, pat: &str) -> Option<usize> where
Self: Sized,

Returns the byte index of the first character of this string slice that matches the pattern. Returns None if the pattern doesn’t match.

The comparison is accent insensitive.

Example

use str_utils::StrUtilsExt;

let s = "Löwe 老虎 Léopard";

assert_eq!(s.find_ai("L"), Some(0));
assert_eq!(s.find_ai("é"), Some(4));
assert_eq!(s.find_ai("Leopard"), Some(13));

fn find_ai_ci(self, pat: &str) -> Option<usize> where
Self: Sized,

Returns the byte index of the first character of this string slice that matches the pattern. Returns None if the pattern doesn’t match.

The comparison is accent / case insensitive.

Example

use str_utils::StrUtilsExt;

let s = "Löwe 老虎 Léopard";

assert_eq!(s.find_ai_ci("l"), Some(0));
assert_eq!(s.find_ai_ci("É"), Some(4));
assert_eq!(s.find_ai_ci("leopard"), Some(13));

fn find_ci(self, pat: &str) -> Option<usize> where
Self: Sized,

Returns the byte index of the first character of this string slice that matches the pattern. Returns None if the pattern doesn’t match.

The comparison is accent / case insensitive.

Example

use str_utils::StrUtilsExt;

let s = "Löwe 老虎 Léopard";

assert_eq!(s.find_ci("l"), Some(0));
assert_eq!(s.find_ci("E"), Some(4));
assert_eq!(s.find_ci("léopard"), Some(13));

fn starts_with_ai(self, pat: &str) -> bool where
Self: Sized,

Returns true if the given pattern matches a prefix of this string slice. Returns false if it does not.

The comparison is accent insensitive.

Example

use str_utils::StrUtilsExt;

assert!("Café Arabica".starts_with_ai("Cafe"));

fn starts_with_ai_ci(self, pat: &str) -> bool where
Self: Sized,

Returns true if the given pattern matches a prefix of this string slice. Returns false if it does not.

The comparison is accent / case insensitive.

Example

use str_utils::StrUtilsExt;

assert!("Café Arabica".starts_with_ai_ci("cafe"));

fn starts_with_ci(self, pat: &str) -> bool where
Self: Sized,

Returns true if the given pattern matches a prefix of this string slice. Returns false if it does not.

The comparison is case insensitive.

Example

use str_utils::StrUtilsExt;

assert!("Cafe Arabica".starts_with_ci("cafe"));

fn no_accent(self) -> String where
Self: Sized,

Transform into a no accent String.

Example

use str_utils::StrUtilsExt;

assert_eq!("Café Arabica".no_accent(), "Cafe Arabica");

fn no_accent_lowercase(self) -> String where
Self: Sized,

Transform into a no accent lowercase String.

Example

use str_utils::StrUtilsExt;

assert_eq!("Café Arabica".no_accent_lowercase(), "cafe arabica");

fn no_accent_uppercase(self) -> String where
Self: Sized,

Transform into a no accent uppercase String.

Example

use str_utils::StrUtilsExt;

assert_eq!("Café Arabica".no_accent_uppercase(), "CAFE ARABICA");

Provided methods

fn contains_ai(self, pat: &str) -> bool where
Self: Sized,

Returns true if the given pattern matches a sub-slice of this string slice. Returns false if it does not.

The comparison is accent insensitive.

Example

use str_utils::StrUtilsExt;

assert!("Café Arabica".contains_ai("Cafe"));

fn contains_ai_ci(self, pat: &str) -> bool where
Self: Sized,

Returns true if the given pattern matches a sub-slice of this string slice. Returns false if it does not.

The comparison is accent / case insensitive.

Example

use str_utils::StrUtilsExt;

assert!("Café Arabica".contains_ai_ci("cafe"));

fn contains_ci(self, pat: &str) -> bool where
Self: Sized,

Returns true if the given pattern matches a sub-slice of this string slice. Returns false if it does not.

The comparison is case insensitive.

Example

use str_utils::StrUtilsExt;

assert!("Café Arabica".contains_ci("café"));